fix(tree): organize exports following alert component pattern
